Change of convention
Separate ground state solver out of time depentent-classes, by removing
method `compute_ground_state`.

A ground state solver now has to be run explicitly, and the
time-dependent solver is initialized with either amplitudes or with a
`cc`-object through the `set_initial_conditions` method. This method
would now become the main method for defining the transformed and/or
truncated basis set(the primed basis set), while the original basis set
is stored in the quantum system (`self.system`). This new basis set is
used for most calculations except in cases where one needs the old
basis set, such as in the q_space equations of oatdccd.